+++ App/xcbuddykit/Sources/Commands/GenerateCommand.swift
@@ -2,16 +2,6 @@ import Basic
 import Foundation
 import Utility
 
-enum GenerateCommandError: Error, ErrorStringConvertible, Equatable {
-    static func == (_: GenerateCommandError, _: GenerateCommandError) -> Bool {
-        return true
-    }
-
-    var errorDescription: String {
-        return ""
-    }
-}
-
 public class GenerateCommand: NSObject, Command {
     /// Command name (static).
     public static let command = "generate"
@@ -33,6 +23,9 @@ public class GenerateCommand: NSObject, Command {
 
     /// Path argument.
     let pathArgument: OptionArgument<String>
+
+    /// Config argument.
+    let configArgument: OptionArgument<String>
 
     /// Initializes the generate command with the argument parser.
     ///
@@ -66,8 +59,13 @@ public class GenerateCommand: NSObject, Command {
         pathArgument = subParser.add(option: "--path",
                                      shortName: "-p",
                                      kind: String.self,
-                                     usage: "The path where the Project.swift file will be generated",
+                                     usage: "The path where the Project.swift file will be generated.",
                                      completion: .filename)
+        configArgument = subParser.add(option: "--config",
+                                       shortName: "-c",
+                                       kind: String.self,
+                                       usage: "The configuration that will be generated.",
+                                       completion: .filename)
     }
 
     /// Runs the command.
@@ -76,13 +74,42 @@ public class GenerateCommand: NSObject, Command {
     /// - Throws: an error if the command cannot be executed.
     public func run(with arguments: ArgumentParser.Result) {
         context.errorHandler.try {
-            var path: AbsolutePath! = arguments.get(pathArgument).map({ AbsolutePath($0) })
-            if path == nil {
-                path = AbsolutePath.current
-            }
+            let path = parsePath(arguments: arguments)
+            let config = parseConfig(arguments: arguments)
             let context = try GeneratorContext(graph: graphLoader.load(path: path))
-            try workspaceGenerator.generate(path: path, context: context)
+            try workspaceGenerator.generate(path: path, context: context, options: GenerationOptions(buildConfiguration: config))
             self.context.printer.print(section: "Generate command succeeded 🎉")
         }
+    }
+
+    /// Parses the arguments and returns the path to the folder where the manifest file is.
+    ///
+    /// - Parameter arguments: argument parser result.
+    /// - Returns: the path to th efolder where the manifest is.
+    private func parsePath(arguments: ArgumentParser.Result) -> AbsolutePath {
+        var path: AbsolutePath! = arguments.get(pathArgument).map({ AbsolutePath($0) })
+        if path == nil {
+            path = AbsolutePath.current
+        }
+        return path
+    }
+
+    /// Parses the arguments and returns the build configuration that we should use for generating the projects.
+    ///
+    /// - Parameters:
+    ///     - arguments: argument parser result.
+    /// - Returns: The build configuration.
+    private func parseConfig(arguments: ArgumentParser.Result) -> BuildConfiguration {
+        var config: BuildConfiguration = .debug
+        if let configString = arguments.get(configArgument) {
+            guard let buildConfiguration = BuildConfiguration(rawValue: configString.lowercased()) else {
+                let error = ArgumentParserError.invalidValue(argument: "config",
+                                                             error: ArgumentConversionError.custom("config can only be debug or release"))
+                context.errorHandler.fatal(error: FatalError.abort(error))
+                return .debug
+            }
+            config = buildConfiguration
+        }
+        return config
     }
 }