+++ App/xcbuddykit/Sources/Generator/ConfigGenerator.swift
@@ -12,22 +12,25 @@ protocol ConfigGenerating: AnyObject {
     ///   - groups: Project groups.
     ///   - sourceRootPath: path to the folder that contains the generated project.
     ///   - context: generation context.
+    ///   - options: generation options.
     /// - Returns: the configuration list reference.
     /// - Throws: an error if the generation fails.
     func generateProjectConfig(project: Project,
                                pbxproj: PBXProj,
                                groups: ProjectGroups,
                                sourceRootPath: AbsolutePath,
-                               context: GeneratorContexting) throws -> PBXObjectReference
+                               context: GeneratorContexting,
+                               options: GenerationOptions) throws -> PBXObjectReference
 
     /// Generates the manifests target configuration.
     ///
     /// - Parameters:
     ///   - pbxproj: Xcode project PBXProj object.
     ///   - context: generation context.
+    ///   - options: generation options.
     /// - Returns: the configuration list reference.
     /// - Throws: an error if the generation fails.
-    func generateManifestsConfig(pbxproj: PBXProj, context: GeneratorContexting) throws -> PBXObjectReference
+    func generateManifestsConfig(pbxproj: PBXProj, context: GeneratorContexting, options: GenerationOptions) throws -> PBXObjectReference
 
     /// Generates the target configuration list and configurations.
     ///
@@ -37,13 +40,15 @@ protocol ConfigGenerating: AnyObject {
     ///   - groups: Project groups.
     ///   - sourceRootPath: path to the folder that contains the generated project.
     ///   - context: generation context.
+    ///   - options: generation options.
     /// - Returns: the configuration list reference.
     /// - Throws: an error if the generation fails.
     func generateTargetConfig(target: Target,
                               pbxproj: PBXProj,
                               groups: ProjectGroups,
                               sourceRootPath: AbsolutePath,
-                              context: GeneratorContexting) throws -> PBXObjectReference
+                              context: GeneratorContexting,
+                              options: GenerationOptions) throws -> PBXObjectReference
 }
 
 /// Config generator.
@@ -66,62 +71,84 @@ final class ConfigGenerator: ConfigGenerating {
     ///   - groups: Project groups.
     ///   - sourceRootPath: path to the folder that contains the generated project.
     ///   - context: generation context.
+    ///   - options: generation options.
     /// - Returns: the confniguration list reference.
     /// - Throws: an error if the generation fails.
     func generateProjectConfig(project: Project,
                                pbxproj: PBXProj,
                                groups: ProjectGroups,
                                sourceRootPath: AbsolutePath,
-                               context: GeneratorContexting) throws -> PBXObjectReference {
-        /// Default build settings
-        let defaultAll = BuildSettingsProvider.projectDefault(variant: .all)
-        let defaultDebug = BuildSettingsProvider.projectDefault(variant: .debug)
-        let defaultRelease = BuildSettingsProvider.projectDefault(variant: .release)
-
-        // Debug
-        var debugSettings: [String: Any] = [:]
-        extend(buildSettings: &debugSettings, with: defaultAll)
-        extend(buildSettings: &debugSettings, with: project.settings?.base ?? [:])
-        extend(buildSettings: &debugSettings, with: defaultDebug)
-        let debugConfiguration = XCBuildConfiguration(name: "Debug")
-        if let debugConfig = project.settings?.debug {
-            extend(buildSettings: &debugSettings, with: debugConfig.settings)
-            if let xcconfigDebug = debugConfig.xcconfig {
-                let fileReference = try fileGenerator.generateFile(path: xcconfigDebug,
-                                                                   in: groups.projectConfigurations(),
-                                                                   sourceRootPath: sourceRootPath,
-                                                                   context: context)
-                debugConfiguration.baseConfigurationReference = fileReference.reference
-            }
-        }
-        debugConfiguration.buildSettings = debugSettings
-        let debugConfigurationReference = pbxproj.objects.addObject(debugConfiguration)
-
-        // Release
-        var releaseSettings: [String: Any] = [:]
-        extend(buildSettings: &releaseSettings, with: defaultAll)
-        extend(buildSettings: &releaseSettings, with: project.settings?.base ?? [:])
-        extend(buildSettings: &releaseSettings, with: defaultRelease)
-        let releaseConfiguration = XCBuildConfiguration(name: "Release")
-        if let releaseConfig = project.settings?.release {
-            extend(buildSettings: &releaseSettings, with: releaseConfig.settings)
-            if let xcconfigRelease = releaseConfig.xcconfig {
-                let fileReference = try fileGenerator.generateFile(path: xcconfigRelease,
-                                                                   in: groups.projectConfigurations(),
-                                                                   sourceRootPath: sourceRootPath,
-                                                                   context: context)
-                releaseConfiguration.baseConfigurationReference = fileReference.reference
-            }
-        }
-        releaseConfiguration.buildSettings = releaseSettings
-        let releaseConfigurationReference = pbxproj.objects.addObject(releaseConfiguration)
-
+                               context: GeneratorContexting,
+                               options: GenerationOptions) throws -> PBXObjectReference {
         /// Configuration list
         let configurationList = XCConfigurationList(buildConfigurations: [])
         let configurationListReference = pbxproj.objects.addObject(configurationList)
-        configurationList.buildConfigurations.append(debugConfigurationReference)
-        configurationList.buildConfigurations.append(releaseConfigurationReference)
+
+        if options.buildConfiguration == .debug {
+            try generateProjectSettingsFor(buildConfiguration: .debug,
+                                           configuration: project.settings?.debug,
+                                           project: project,
+                                           groups: groups,
+                                           sourceRootPath: sourceRootPath,
+                                           context: context,
+                                           pbxproj: pbxproj,
+                                           configurationList: configurationList)
+        } else if options.buildConfiguration == .release {
+            try generateProjectSettingsFor(buildConfiguration: .release,
+                                           configuration: project.settings?.release,
+                                           project: project,
+                                           groups: groups,
+                                           sourceRootPath: sourceRootPath,
+                                           context: context,
+                                           pbxproj: pbxproj,
+                                           configurationList: configurationList)
+        }
         return configurationListReference
+    }
+
+    /// Generate the project settings for a given configuration (e.g. Debug or Release)
+    ///
+    /// - Parameters:
+    ///   - buildConfiguration: build configuration (e.g. Debug or Release)
+    ///   - configuration: configuration from the project specification.
+    ///   - project: projec specification.
+    ///   - groups: project groups.
+    ///   - sourceRootPath: path that points to the folder where the project is being created.
+    ///   - context: generation context.
+    ///   - pbxproj: PBXProj object.
+    ///   - configurationList: configurations list.
+    /// - Throws: an error if the generation fails.
+    private func generateProjectSettingsFor(buildConfiguration: BuildConfiguration,
+                                            configuration: Configuration?,
+                                            project: Project,
+                                            groups: ProjectGroups,
+                                            sourceRootPath: AbsolutePath,
+                                            context: GeneratorContexting,
+                                            pbxproj: PBXProj,
+                                            configurationList: XCConfigurationList) throws {
+        let variant: BuildSettingsProvider.Variant = (buildConfiguration == .debug) ? .debug : .release
+        let defaultConfigSettings = BuildSettingsProvider.projectDefault(variant: variant)
+        let defaultSettingsAll = BuildSettingsProvider.projectDefault(variant: .all)
+
+        var settings: [String: Any] = [:]
+        extend(buildSettings: &settings, with: defaultSettingsAll)
+        extend(buildSettings: &settings, with: project.settings?.base ?? [:])
+        extend(buildSettings: &settings, with: defaultConfigSettings)
+
+        let variantBuildConfiguration = XCBuildConfiguration(name: buildConfiguration.rawValue.capitalized)
+        if let variantConfig = configuration {
+            extend(buildSettings: &settings, with: variantConfig.settings)
+            if let xcconfig = variantConfig.xcconfig {
+                let fileReference = try fileGenerator.generateFile(path: xcconfig,
+                                                                   in: groups.projectConfigurations(),
+                                                                   sourceRootPath: sourceRootPath,
+                                                                   context: context)
+                variantBuildConfiguration.baseConfigurationReference = fileReference.reference
+            }
+        }
+        variantBuildConfiguration.buildSettings = settings
+        let debugConfigurationReference = pbxproj.objects.addObject(variantBuildConfiguration)
+        configurationList.buildConfigurations.append(debugConfigurationReference)
     }
 
     /// Generates the configuration for the manifests target.
@@ -129,18 +156,11 @@ final class ConfigGenerator: ConfigGenerating {
     /// - Parameters:
     ///   - pbxproj: PBXProj object.
     ///   - context: generation context.
+    ///   - options: generation options.
     /// - Returns: the configuration list object reference.
     /// - Throws: an error if the genreation fails.
-    func generateManifestsConfig(pbxproj: PBXProj, context: GeneratorContexting) throws -> PBXObjectReference {
+    func generateManifestsConfig(pbxproj: PBXProj, context: GeneratorContexting, options: GenerationOptions) throws -> PBXObjectReference {
         let configurationList = XCConfigurationList(buildConfigurations: [])
-        let debugConfig = XCBuildConfiguration(name: "Debug")
-        let debugConfigReference = pbxproj.objects.addObject(debugConfig)
-        debugConfig.buildSettings = BuildSettingsProvider.targetDefault(variant: .debug, platform: .macOS, product: .framework, swift: true)
-        let releaseConfig = XCBuildConfiguration(name: "Release")
-        let releaseConfigReference = pbxproj.objects.addObject(releaseConfig)
-        releaseConfig.buildSettings = BuildSettingsProvider.targetDefault(variant: .release, platform: .macOS, product: .framework, swift: true)
-        configurationList.buildConfigurations.append(debugConfigReference)
-        configurationList.buildConfigurations.append(releaseConfigReference)
         let configurationListReference = pbxproj.objects.addObject(configurationList)
 
         let addSettings: (XCBuildConfiguration) throws -> Void = { configuration in
@@ -148,8 +168,20 @@ final class ConfigGenerator: ConfigGenerating {
             configuration.buildSettings["FRAMEWORK_SEARCH_PATHS"] = frameworkParentDirectory.asString
             configuration.buildSettings["SWIFT_VERSION"] = Constants.swiftVersion
         }
-        try addSettings(debugConfig)
-        try addSettings(releaseConfig)
+        if options.buildConfiguration == .debug {
+            let debugConfig = XCBuildConfiguration(name: "Debug")
+            let debugConfigReference = pbxproj.objects.addObject(debugConfig)
+            debugConfig.buildSettings = BuildSettingsProvider.targetDefault(variant: .debug, platform: .macOS, product: .framework, swift: true)
+            configurationList.buildConfigurations.append(debugConfigReference)
+            try addSettings(debugConfig)
+        }
+        if options.buildConfiguration == .release {
+            let releaseConfig = XCBuildConfiguration(name: "Release")
+            let releaseConfigReference = pbxproj.objects.addObject(releaseConfig)
+            releaseConfig.buildSettings = BuildSettingsProvider.targetDefault(variant: .release, platform: .macOS, product: .framework, swift: true)
+            configurationList.buildConfigurations.append(releaseConfigReference)
+            try addSettings(releaseConfig)
+        }
         return configurationListReference
     }
 
@@ -161,13 +193,15 @@ final class ConfigGenerator: ConfigGenerating {
     ///   - groups: Project groups.
     ///   - sourceRootPath: path to the folder that contains the generated project.
     ///   - context: generation context.
+    ///   - options: generation options.
     /// - Returns: the configuration list reference.
     /// - Throws: an error if the generation fails.
     func generateTargetConfig(target _: Target,
                               pbxproj: PBXProj,
                               groups _: ProjectGroups,
                               sourceRootPath _: AbsolutePath,
-                              context _: GeneratorContexting) throws -> PBXObjectReference {
+                              context _: GeneratorContexting,
+                              options: GenerationOptions) throws -> PBXObjectReference {
         let configurationList = XCConfigurationList(buildConfigurations: [])
         let configurationListReference = pbxproj.objects.addObject(configurationList)
         // TODO: