+++ App/xcbuddykit/Tests/Generator/Mocks/MockProjectGenerator.swift
@@ -3,7 +3,7 @@ import Foundation
 @testable import xcbuddykit
 
 final class MockProjectGenerator: ProjectGenerating {
-    func generate(project _: Project, sourceRootPath _: AbsolutePath?, context _: GeneratorContexting) throws -> AbsolutePath {
+    func generate(project _: Project, sourceRootPath _: AbsolutePath?, context _: GeneratorContexting, options _: GenerationOptions) throws -> AbsolutePath {
         return AbsolutePath("/")
     }
 }